requirements.txt
# OSINT-Nexus Dependencies
# Core GUI
PyQt6>=6.4.0
pyqtgraph>=0.13.0

# Graph Analysis
networkx>=3.0
scipy>=1.10.0

# Async & HTTP
aiohttp>=3.8.0
requests>=2.28.0

# OSINT Libraries
phonenumbers>=8.13.0
python-whois>=0.8.0
dnspython>=2.3.0

# Web Scraping
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Optional OSINT APIs (Free tiers)
shodan>=1.28.0

# Utilities
python-dotenv>=1.0.0

# Advanced Features
reportlab>=4.0.0     # PDF report generation
Pillow>=9.0.0        # Image forensics (EXIF extraction)

